pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "indot-ec-agent"
version = "0.1.0"
description = "INDOT-first erosion control recommendation agent"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "INDOT EC Agent"}]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.3.0",
  "numpy>=1.26.0",
  "scipy>=1.11.0",
  "PyPDF2>=3.0.0",
]

[project.scripts]
ec_agent = "ec_agent.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
